Change the catalog's upgrade graph strategy to semver-skip patch
Default strategy "replaces" requires to manually set the replaces field in the
bundle's csv file. Using the "semver-skippatch" strategy allows olm to upgrade the
operator following semver, with the particularity that it will skip going through all
patch versions for a minor release and directly go to the latest available.