Hardcode the name of the container for all controllers
The name of the containers within the Pod should not be dependant on the
name of the component (which changes in the case of canary Deployments
for example). This is important as the shutdown manager features in
marin3r depend on the name of the containers being always the same.