Implement fix for https://github.com/redhat-cop/operator-utils/pull/71
saas-operator managed Services relied on a behavior from operator-utils
where each time a resource controller was restarted, an update operation
was performed on the managed resource. This behavior was problematic in
several scenarios and upstream operator-utils made a change to address
these problems by always performing patch operations instead of updates.
This causes the side effect that strategic merge patches (the ones being
used) cannot delete keys in hash maps, as maps are always merged. This
breaks our implementation of the canary Deployments, which relies on
adding/removing labels to the Service selector to send/not send traffic
to the canary Deployment.

The change in operator-utils is https://github.com/redhat-cop/operator-utils/pull/71

The fix for this change in behavior is provide the basereconciler with
the hability to use a custom reconciler. When a resource type within the
basereconciler implements the ResourceWithCustomReconciler interface,
the ResourceReconciler() function is called for that specific resource
type instead of sending the resource to the default operator-utils
reconcile logic. A custom reconciler has been implemented for the
Service resources.