Simplify the workload reconciler
The workload reconciler was originally conceived to deploy several
workloads behind a traffic manager. This workeed well but is too complex
for what we are using it for, which is just a main workload with a
service and an optional canary workload. This commit modifies the model
to adapt it to how we are actually using it so the code reads more
naturally.