Refactor again to use a map in the external API
There is no simple way of making kustomize work with strategic merge
patches in custom resources so this commit changes the external API to
use a map of envoy dynamic configurations instead of a list. The API has
methods to transform the map into the list that the internal API
expects.