Avoid failover delays in twemproxy reconfig when using master targets
It seems that even though the "sentinel master <shard_name>" command
returns updated information about the master as soon as a slave is
promoted, there is an exeption with the sentinel instance that acts as
"failover leader". The failover leader is the instance that actually
performs the shard reconfigurations, and in this case, its "sentinel
master <shard_name>" command only returns updated information when all
the slaves have been reconfigured to point to the new master. This
causes delays in twemproxy reconfiguration if this is the instance being
used to "discover" the shard.
To detect this situation, add a step that checks the master address with
the command "sentinel get-master-addr-by-name <shard_name>". This
command always returns updated master information, even if we are
querying the leader sentinel instance.