Implement Replace mode for publishingStrategies
To unlock the possibility of deploying 2 different publishing strategies
for the same endpoint, the whole list of strategies needs to be
replaces, because with Merge only one strategy can exists per endpoint,
otherwise it wouldn't know how to do the merge.
Replace allows exposing the same endpoint through several different
strategies in parallel, which unlocks the possiblity of DNS based migrations without
downtime.